+++ CHANGELOG.md
@@ -31,6 +31,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     * from `pytorch_lightning.metrics.functional.reduction` removed `reduce`, `class_reduce`	
 
 
+- Removed deprecated `ModelCheckpoint` arguments `prefix`
+
+
 ### Fixed
 
 - Made the `Plugin.reduce` method more consistent across all Plugins to reflect a mean-reduction by default ([#6011](https://github.com/PyTorchLightning/pytorch-lightning/pull/6011))